ssize_t instead of intmax_t

+++ inc/utils.h
@@ -15,14 +15,14 @@
 
 # include "ft_printf.h"
 
-uintmax_t	manage_length_unsigned(uintmax_t val, t_length length);
-intmax_t	manage_length_signed(intmax_t val, t_length length);
-char		*convert_base(uintmax_t unsgnd,
+size_t	manage_length_unsigned(size_t val, t_length length);
+ssize_t	manage_length_signed(ssize_t val, t_length length);
+char		*convert_base(size_t unsgnd,
 	int base, t_specifier specifier, bool is_unsigned);
 char		*ft_strndup(const char *src, size_t size);
 bool		check_null_value_and_prec(int prec, char **result);
-size_t		get_numlen(uintmax_t usigned_num, short base, bool is_unsigned);
-void		print_num(uintmax_t unsigned_num, bool is_unsigned);
+size_t		get_numlen(size_t usigned_num, short base, bool is_unsigned);
+void		print_num(size_t unsigned_num, bool is_unsigned);
 unsigned int count_bits(unsigned int value);
 
 int cmp_len(t_length curr, t_length new);
