+++ src/manage_hex.c
@@ -15,11 +15,11 @@
 int	print_hex(t_handler *handler, va_list args)
 {
 	int			chars_printed;
-	uintmax_t	value;
+	size_t	value;
 	char		*result;
 	size_t		len;
 
-	value = manage_length_unsigned(va_arg(args, uintmax_t), handler->length);
+	value = manage_length_unsigned(va_arg(args, size_t), handler->length);
 	if (!value && handler->specifier != POINTER)
 		handler->flags.hash = value != 0 && handler->flags.hash;
 	if (handler->specifier == POINTER)
